feat: 🆕 empty string to null for Nullable<>
